[skip ci]
• pytorchcomponents.py
  remove pytorch_function_creator

• function.py
  Function: add _get_pytorch_fct_param_value()

• transferfunctions.py,
  combinationfunctions.py
  learningfunctions.py:
  - add _gen_pytorch_fct()
